requirements.txt
certifi==2024.7.4
confluent-kafka==2.4.0
elasticsearch7==7.17.9
python-dotenv==1.0.1
urllib3==1.26.19
